# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bluebird B2B Enterprise Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [available, booked, maintenance]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/available:
    get:
      operationId: listAvailableVehicles
      tags: [vehicles]
      summary: List vehicles available between two dates (excludes booked & maintenance)
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleDetail"
    patch:
      operationId: updateVehicleStatus
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, active, completed, cancelled]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        plateNumber:
          type: string
        category:
          type: string
          description: "e.g. Goldenbird Sedan, Goldenbird SUV, Bigbird Bus"
        model:
          type: string
        year:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, booked, maintenance]
        maintenanceCompletionDate:
          type: string
          format: date
          nullable: true
        maintenanceNote:
          type: string
          nullable: true
        dailyRate:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, plateNumber, category, model, year, capacity, status, dailyRate, createdAt]
    VehicleDetail:
      allOf:
        - $ref: "#/components/schemas/Vehicle"
        - type: object
          properties:
            orders:
              type: array
              items:
                $ref: "#/components/schemas/Order"
          required: [orders]
    CreateVehicleBody:
      type: object
      properties:
        plateNumber:
          type: string
        category:
          type: string
        model:
          type: string
        year:
          type: integer
        capacity:
          type: integer
        dailyRate:
          type: number
      required: [plateNumber, category, model, year, capacity, dailyRate]
    UpdateVehicleStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [available, booked, maintenance]
        maintenanceCompletionDate:
          type: string
          format: date
          nullable: true
        maintenanceNote:
          type: string
          nullable: true
      required: [status]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        vehicleId:
          type: integer
        vehiclePlate:
          type: string
        driverId:
          type: integer
        driverName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        pickupLocation:
          type: string
        dropoffLocation:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [draft, active, completed, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, orderNumber, clientId, clientName, vehicleId, vehiclePlate, driverId, driverName, startDate, endDate, pickupLocation, dropoffLocation, price, status, createdAt]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            invoice:
              nullable: true
              $ref: "#/components/schemas/Invoice"
    CreateOrderBody:
      type: object
      properties:
        clientId:
          type: integer
        vehicleId:
          type: integer
        driverId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        pickupLocation:
          type: string
        dropoffLocation:
          type: string
        price:
          type: number
        notes:
          type: string
          nullable: true
      required: [clientId, vehicleId, driverId, startDate, endDate, pickupLocation, dropoffLocation, price]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, active, completed, cancelled]
      required: [status]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        orderId:
          type: integer
        orderNumber:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, outstanding]
        issuedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        paidDate:
          type: string
          format: date
          nullable: true
      required: [id, invoiceNumber, orderId, orderNumber, clientId, clientName, amount, status, issuedDate, dueDate]
...