# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluebird B2B Enterprise Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  clientId: number;
  vehicleId: number;
  driverId: number;
  startDate: Date;
  endDate: Date;
  pickupLocation: string;
  dropoffLocation: string;
  price: number;
  notes?: string | null;
}