# lib/api-zod/src/generated/types/createVehicleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluebird B2B Enterprise Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVehicleBody {
  plateNumber: string;
  category: string;
  model: string;
  year: number;
  capacity: number;
  dailyRate: number;
}