# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluebird B2B Enterprise Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleStatus } from "./vehicleStatus";

export interface Vehicle {
  id: number;
  plateNumber: string;
  /** e.g. Goldenbird Sedan, Goldenbird SUV, Bigbird Bus */
  category: string;
  model: string;
  year: number;
  capacity: number;
  status: VehicleStatus;
  maintenanceCompletionDate?: Date | null;
  maintenanceNote?: string | null;
  dailyRate: number;
  createdAt: Date;
}